chore: #55: change static cast from in to uint

+++ src/application/features/get-to-do-by-id/get-to-do-by-id-handler.cpp
@@ -3,14 +3,14 @@
 static ToDoDTO mapToDTO(const ToDo& todo)
 {
     ToDoDTO dto;
-    dto.id = static_cast<int>(todo.id());
+    dto.id = static_cast<std::uint64_t>(todo.id());
     dto.name = todo.name();
     return dto;
 }
 
 GetToDoByIdResponse GetToDoByIdHandler::handle(std::uint64_t id)
 {
-    auto todo = _getToDoByIdQuery.get(static_cast<int>(id));
+    auto todo = _getToDoByIdQuery.get(static_cast<std::uint64_t>(id));
 
     if (!todo)
         return { std::nullopt };
