+++ src/application/features/hard-delete-to-do/hard-delete-to-do-handler.cpp
@@ -3,6 +3,6 @@
 // TODO(https://github.com/TourmalineCore/to-dos-api-cpp/issues/38): add here a check for not found todo
 bool HardDeleteToDoHandler::handle(std::uint64_t id)
 {
-    (void) _hardDeleteToDoCommand.execute(static_cast<int>(id));
+    (void) _hardDeleteToDoCommand.execute(static_cast<std::uint64_t>(id));
     return true;
 }
