+++ src/application/features/soft-delete-to-do/soft-delete-to-do-handler.cpp
@@ -3,6 +3,6 @@
 // TODO(https://github.com/TourmalineCore/to-dos-api-cpp/issues/38): add here a check for not found todo
 SoftDeleteToDoResponse SoftDeleteToDoHandler::handle(std::uint64_t id)
 {
-    const std::uint64_t deletedId = _softDeleteToDoCommand.execute(static_cast<int>(id));
+    const std::uint64_t deletedId = _softDeleteToDoCommand.execute(static_cast<std::uint64_t>(id));
     return { deletedId };
 }
